Remove dead pipelines/schemas and centralize publication-term ID hash
