+++ src/bioetl/application/core/publication_term_data_source.py
@@ -22,9 +22,11 @@ Architecture:
 
 from __future__ import annotations
 
-import hashlib
 from typing import TYPE_CHECKING, Any, Self
 
+from bioetl.application.core.publication_term_utils import (
+    compute_publication_term_entity_id,
+)
 from bioetl.domain.ports import FilterableDataSourcePort
 
 if TYPE_CHECKING:
@@ -326,9 +328,11 @@ class PublicationTermDataSource:
             Entity ID string (first 16 chars of SHA256 hex digest).
 
         """
-        normalized_term = term.lower().strip() if term else ""
-        composite = f"{document_chembl_id}:{term_type}:{normalized_term}"
-        return hashlib.sha256(composite.encode()).hexdigest()[:16]
+        return compute_publication_term_entity_id(
+            document_chembl_id=document_chembl_id,
+            term_type=term_type,
+            term=term,
+        )
 
     async def health_check(self) -> HealthStatus:
         """Delegate health check to wrapped adapter."""
