--- src/bioetl/application/pipelines/pubmed/publication.py
@@ -1,18 +1,0 @@
-# src/bioetl/application/pipelines/pubmed/publication.py
-"""PubMed Publication Pipeline.
-
-Transformer is injected via DI from GenericPipelineFactory (REQ-ARCH-DI-007).
-"""
-
-from __future__ import annotations
-
-from bioetl.application.core.base import BasePipeline
-
-
-class PubMedPublicationPipeline(BasePipeline):
-    """Пайплайн для данных о публикациях из PubMed.
-
-    Transformer is injected via DI from GenericPipelineFactory.
-    """
-
-    # transform_bronze_to_silver() is inherited from BasePipeline
