+++ src/bioetl/composition/services/versioning.py
@@ -159,49 +159,3 @@ def get_pipeline_version(
         return pkg_version("bioetl")
     except Exception:
         return "unknown"
-
-
-def get_full_git_commit() -> str | None:
-    """Get the full git commit hash (40 characters).
-
-    Similar to get_git_commit() but returns the full hash instead of
-    the short version. Useful for exact reproducibility tracking.
-
-    Returns:
-        Full git commit hash (e.g., 'abc1234...') or None.
-    """
-    try:
-        result = subprocess.run(
-            ["git", "rev-parse", "HEAD"],
-            capture_output=True,
-            text=True,
-            timeout=5,
-            check=False,
-        )
-        if result.returncode == 0:
-            return result.stdout.strip()
-        return None
-    except (subprocess.SubprocessError, FileNotFoundError, OSError):
-        return None
-
-
-def is_git_dirty() -> bool:
-    """Check if the git working directory has uncommitted changes.
-
-    Returns:
-        True if there are uncommitted changes, False otherwise.
-        Returns False if not in a git repository or git is unavailable.
-    """
-    try:
-        result = subprocess.run(
-            ["git", "status", "--porcelain"],
-            capture_output=True,
-            text=True,
-            timeout=5,
-            check=False,
-        )
-        if result.returncode == 0:
-            return bool(result.stdout.strip())
-        return False
-    except (subprocess.SubprocessError, FileNotFoundError, OSError):
-        return False
