+++ tests/unit/application/pipelines/test_pipeline_registrations.py
@@ -24,15 +24,21 @@ _PIPELINE_MODULES: list[tuple[str, str]] = [
     ("bioetl.application.pipelines.chembl._pipelines", "ChEMBLMoleculePipeline"),
     ("bioetl.application.pipelines.chembl._pipelines", "ChEMBLProteinClassPipeline"),
     ("bioetl.application.pipelines.chembl._pipelines", "ChEMBLPublicationPipeline"),
-    ("bioetl.application.pipelines.chembl._pipelines", "ChEMBLPublicationSimilarityPipeline"),
+    (
+        "bioetl.application.pipelines.chembl._pipelines",
+        "ChEMBLPublicationSimilarityPipeline",
+    ),
     ("bioetl.application.pipelines.chembl._pipelines", "ChEMBLPublicationTermPipeline"),
-    ("bioetl.application.pipelines.chembl._pipelines", "ChEMBLSubcellularFractionPipeline"),
+    (
+        "bioetl.application.pipelines.chembl._pipelines",
+        "ChEMBLSubcellularFractionPipeline",
+    ),
     ("bioetl.application.pipelines.chembl._pipelines", "ChEMBLTargetPipeline"),
     ("bioetl.application.pipelines.chembl._pipelines", "ChEMBLTargetComponentPipeline"),
     ("bioetl.application.pipelines.chembl._pipelines", "ChEMBLTissuePipeline"),
-    ("bioetl.application.pipelines.pubchem.compound", "PubChemCompoundPipeline"),
-    ("bioetl.application.pipelines.pubmed.publication", "PubMedPublicationPipeline"),
-    ("bioetl.application.pipelines.uniprot.protein", "UniProtProteinPipeline"),
+    ("bioetl.application.pipelines.pubchem", "PubChemCompoundPipeline"),
+    ("bioetl.application.pipelines.pubmed", "PubMedPublicationPipeline"),
+    ("bioetl.application.pipelines.uniprot", "UniProtProteinPipeline"),
 ]
